DynamoDB implemented

+++ src/MCGAssignment.TodoList.Core/Enums/TaskPriority.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 namespace MCGAssignment.TodoList.Core.Enums;
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
-public enum TaskPriority : int
+public enum TaskPriority : byte
 {
     Low,
     Normal,
